va_tpi: vaCreateSurfaceFromKBuf for streaming the video playback to remote monitor
For such use case, the pipeline is: decode->encode->send to network.
The local decode/playback and the encode process are seperated, thus it
needs the buffer sharing b/w the decode and encode in seperate process.
This API creates/wraps the decoded surface into an encode surface. Its
input parameter is the kernel buffer handle of decoded surface which is
only understandable by driver implementation.

Change-Id: I7953ecf3b7333c1cdb6af5a0f45b054c8840f8e2
Signed-off-by: jiguoliang <guoliangx.ji@intel.com>
Signed-off-by: Austin Yuan <shengquan.yuan@gmail.com>