API: h264 decode: introduce num_emulation_prevention_bytes.
num_emulation_prevention_bytes represent the number of emulation prevention
bytes available in the slice_header(). This helps VA drivers determine how
many emulation prevention bytes there are so that to program the hardware
appropriately, without having to parse the slice_header() again.

Rationale: it is not always possible to parse the slice_header() again.
For instance, if "protected slices" are used, the slice data is uploaded
into a VAProtectedSliceDataBuffer and it is no longer possible to read
bytes from it afterwards. e.g. it could be in encrypted memory region
or protected memory region only accessible from the video decoder.

In newly written code, this member shall be filled in appropriately along
with emulation_prevention_bytes_flag set to 1.

Signed-off-by: Gwenole Beauchesne <gwenole.beauchesne@intel.com>