jpeg: mark default Huffman table as valid.
When no Huffman table is defined in the JPEG segments, then the default
Huffman tables are used, but values were not marked as valid. Fixed this
to notify HW decoder that it needs to load the default tables.

Note: this is not specifically needed to be handled at the application
level. The driver also ought to generate default tables, should the HW
decoder require so and no table was provided from the application.

Signed-off-by: Gwenole Beauchesne <gwenole.beauchesne@intel.com>