Libva: Add range parameter flag to reflect the JPEG/JFIF CSC coeff matrix
V3->V4: Update the range flag bit definition to avoid the conflict with
the AUTO_NOISEREDUCTION filter flag.

V2->V3: Follow Penne to describe the purpose of reduced range and full range

V1->V2: Follow Penne and Jonathan to reduce the JPEG flag. Instead it is
derived from the combination of the range flag and BT601 CSC standard.
At the same time the chroma_siting flag in va_vpp.h is rename as
input_surface_flag to hold the corresponding range flag on the input
pixel data.

Signed-off-by: Zhao Yakui <yakui.zhao@intel.com>