va: drop VAEntrypointMax enumeration.
This is an oversight, and that was supposed to be removed earlier.
User applications cannot rely on this value to statically allocate
an array of possible VA entrypoints. The correct interface is to
use vaMaxNumEntrypoints() and dynamically allocate that array.

Rationale: VA entrypoints values are maintained sparsely, so this
can yield to a large number overall.

Signed-off-by: Gwenole Beauchesne <gwenole.beauchesne@intel.com>
(cherry picked from commit 0b71390aac4b0b7971d5c279a87005e4c33d1b45)