Add one API to allow the user to specify the loaded backend driver
Sometimes there exist multiple backend vaapi drivers on one system. In such case
user hopes to load the specific driver based on the request. Currently it can
be specified by using LIBVA_DRIVER_NAME environment. But it is not convenient.
So one API is added to allow the user to specify the loaded backend vaapi driver.
And it is called by using the following sequence:
   >vaGetDisplay(
   >vaSetDriverName(va_dpy, driver_name)
   >vaInitialize

V1->V2: Use the reserved space to store override_driver_name based on Haihao's
suggestion.

Signed-off-by: Zhao Yakui <yakui.zhao@intel.com>