Add the comment for framerate in VAEncMiscParameterFrameRate
To express a frame rate, such as 7.5, we have to use 'numerator / denominator'.
The high 2 bytes will be used as denominator and the low 2 bytes will be used
as numerator.

To keep backward compatible, the low 2 bytes is the real frame rate when the high
2 bytes is 0.

Signed-off-by: Xiang, Haihao <haihao.xiang@intel.com>
Reviewed-by: Sean V Kelley <seanvk@posteo.de>
(cherry picked from commit 1ca62cedd41ed2061c3a9ddafadc59a4692a07c9)