wayland: Do not rely on external library for wl_drm_interface
Doing the runtime lookup is error-prone (as indicated in the code
already by the two library names) and not necessary at all.
Just compile wl_drm_interface as part of libva which is the usual
way to use extra Wayland protocols. Care must be taken as to not
accidentally export the symbol though.