add encoding dirty rectangle. Read-only.
user provided rectangles which indicate the rectangular areas
where the content has changed as compared to the previous picture.  The regions of the
picture that are not covered by dirty rectangles are assumed to have not changed
compared to the previous picture.  The encoder may do some optimizations based on
this information.the ROI information is passed to the driver using VAEncMiscParameterTypeDirtyRect.

Signed-off-by: Carl.Zhang <carl.zhang@intel.com>