Fix issue with static analysis scan
Coverity fails to parse the expanded value of
va_deprecated and causes the scan to abort.  This
problem also affects downstream dependents of
va.h that get scanned with coverity, too.  This
results in only partially-scanned code with the
following coverity errors:

"../../va/va.h", line 361: error #67: expected a "}"
      VAProfileH264Baseline va_deprecated_enum = 5,
                            ^

To fix this, we don't need to define a value for
va_deprecated or va_deprecated_enum during a
coverity scan.

Change-Id: I1e8f2ab4d579ac7e07d4031af97c5ef5603eeb78
Signed-off-by: U. Artie Eoff <ullysses.a.eoff@intel.com>
Signed-off-by: Badiuzzaman Iskhandar <badiuzzaman.azzarfan.bin.iskhandar@intel.com>
Tracked-On: https://jira01.devtools.intel.com/browse/OAM-62660
Reviewed-on: https://android.intel.com:443/629617