add surface hint for recon/reference/raw
this surface hint is used to require usage hint
for some codec , to optimize the memory footprint
reconstruct surface is use another format
for example, if input surface is AYUV for encode, it 444 format,
backend driver maybe just use YUV to be reference, A is ignored
to safe memory. in this scenario, application should not confuse
the raw input surface and reconstruct/reference surface.

Signed-off-by: Carl Zhang <carl.zhang@intel.com>