Fixes file descriptor leak
If code calls va_isDRI2Connected multiple times there is no protection
against calling open(device_name, O_RDWR) again and losing the reference
to the original file descriptor.

A recent change (bc8a12) calls the method twice each time libva is
initialized, thus causing a leak of device file descriptors.

This change moves the initializing of the fd and auth_type to the
creation of the dri_state and then checks to see if they have already
been initialized. If so, the method va_isDRIConnected returns true
and fills in driver_name.