meson: Copy public headers to build directory
In order to use libva in a subproject[1] or in an "uninstalled"
environment, libva needs to fix its public headers layout in its
source directory. A previous fix was proposed[2] which moved
those headers as expected, but it was closed because it was too
invasive.

This new approach is less invasive and oriented to meson subprojects
or uninstalled setups, which has a wider audience.

The fix is two-fold:

1. Bump the required version of meson in order to support this new
   feature. And update deprecated syntax.
2. Copy each public header file into subproject's build directory.

1. https://mesonbuild.com/Subprojects.html
2. https://github.com/intel/libva/pull/433

squashme