Change the default log level from info to errors
libva by default writes log messages to stdout, but the stdouts of GUI
applications are connected to the systemd journal these days.

This results in a lot of unnecessarily repeated log messages of the form
> libva info: VA-API version 1.13.0
> libva info: Trying to open /usr/lib64/dri/radeonsi_drv_video.so
> libva info: Found init function __vaDriverInit_1_13
> libva info: va_openDriver() returns 0
as e.g. Firefox creates and destroys video decoders for page elements.

This patch reduces the default log level from info to errors only.

Signed-off-by: Nicholas Miell <nmiell@gmail.com>