va_trace: va_TraceSurfaceAttributes should check the VASurfaceAttribMemoryType
In the VASurfaceAttribExternalBufferDescriptor docs, it states that the memory type must be referred to
to know which type is passed with this attribute or use the default if not specified.
This change checks if a memory type attibute is passed that does not match the VASurfaceAttribExternalBuffers type.
Without this change, other memory types will cause va_TraceSurfaceAttributes to access garbage memory.

Signed-off-by: Sil Vilerino <sivileri@microsoft.com>