autotools: use wayland-scanner private-code
The "code" option was deprecated with wayland-scanner 1.15. The version
is available even in very old distros Debian Stretch and Ubuntu Bionic
(18.04), so it's fine to depend on it.

Here we replace the code option with private-code, since the code is
meant only for internal consumption.

With that done, we can drop the no longer needed export.c workaround.

Signed-off-by: Emil Velikov <emil.l.velikov@gmail.com>