ci: add meson permutation in the Ubuntu builds
A number of Linux distributions have switched to the meson build. In
particular - Arch, Fedora, Gentoo and more are to follow in due time.

Even the libva builds that Chrome uses (aka instrumented libraries) is
using meson (on top of Ubuntu, currently aiming at switching to Focal).

Note: `meson compile` was introduced with meson 0.54, where we use 0.53
since that's available in Ubuntu 20.04. As such we fallback to calling
ninja directly

v2: group the meson/autotool sections

Signed-off-by: Emil Velikov <emil.l.velikov@gmail.com>