Always build with va-messaging API
Currently the API is enabled in meson, autotools and Android (non-user)
builds.

We don't have any CI to check if the nothing is broken when building w/o
the API. Plus overall disabling the API at build-time breaks the
contract that the users (both end-users and API users) expect.

Namely: the API works and can be controlled via the config file and
environment variable.

Signed-off-by: Emil Velikov <emil.l.velikov@gmail.com>