va/x11/va_nvctrl: remove some dead code
Drop a bunch of set but unused variables and associated dead code.

Note: if we are to remove the version handling, then 2/3 of the code in
this file will become unused. Yet it's fairly subtle code and there's a
notable chance of detection breaking if we remove it.

So keep the variables around and (void) prefix them so the compiler does
not warn.

Signed-off-by: Emil Velikov <emil.l.velikov@gmail.com>