va: add vaGetDriverNames() internal ABI
The current GetDriver API is rather fragile and bonkers.

In particular it does a two-step pass - number vs names. That in itself
is not great, since it assumes the devices do not change in between.
Additionally there is the assumption of index X will match device (name)
X, which does not hold true.

The most prominent issue is the inability to provide accurate number and
list of names.

Instead, we can have a single API, that provides both pieces of data in
one go - shorter and simpler code - in reliable manner.

Signed-off-by: Emil Velikov <emil.l.velikov@gmail.com>