win32: Fix driver loader to support required absolute paths
The installed drivers in the registry via driver store
MUST be absolute paths, so we made some changes to support this:

- Change va_DisplayContextGetDriverNameByIndex to extract the
driver name (minus the drv_video prefix) from the absolute path
in the registry key string.

- Implement vaGetDriverSearchPathByIndex to provide the
search driver absolute path for the registered driver

- Provide a .\ search path for the default vaon12
driver for apps to avoid having to set LIBVA_DRIVERS_PATH
for the default.

Signed-off-by: Sil Vilerino <sivileri@microsoft.com>