Add HEVC RefPicSetLtCurr and NumPocLtCurr picture parameters
Corresponds to the variables of same name as defined in section 8.3.2
of the HEVC spec.

The RefPicSetStCurrBefore and RefPicSetStCurrAfter can be derived
by sorting them, however there is no defined order for the
RefPicSetLtCurr and thus it must be provided explicitly.

This is needed for AMD driver and also to match the DXVA and Vulkan
Video interfaces.

Signed-off-by: David Rosca <david.rosca@amd.com>